Update default usage of InfoList component (#495)
* Update default usage of InfoList component

Per Matt's request, all unordered lists should use a dashed list or a bulleted list. All lists of link should use dashes, all others should use bullets. The unspecified usages of the InfoList component seem to be lists of links, where the specified usages are using the bullet. Thus changing the default to be the dash. Also removing an unnecessary prop from the sibling BulletInfoList. Finally, updating the dash to be a medium length dash instead of the short one that was being used.

* update usage of infolist for a11y pages

* replace `.dash-list`